refactor(alt-text): collapse revalidate hook back into a single file
Drop the .core.ts split, the altTextHealthTags.ts extraction, and the
revalidate hook unit tests. Those existed solely to make after()-deferred
revalidation unit-testable without next/cache installed in the test runner;
the tests locked in implementation details (after() vs queueMicrotask, exact
error messages) rather than user-observable behavior, and the user-observable
behavior — payload.create from onInit not crashing the admin render — needs
a real Next.js runtime to test, which is out of scope here.

The fix itself is unchanged: revalidateTag is still deferred via after()
to escape the render scope, with a fallback to a direct call (and the
existing 'static generation store missing' warn-and-skip) for CLI scopes.
